// oc_prod/RewardsApp/src/app/enums/urlenums.enum.ts
export enum UrlEnums {
    BASE_URL = 'https://app.optculture.com/subscriber',
    // CMS_BASE_URL = 'https://b-staging.optculture.cloud',

    LOGIN_URL = '/loyaltyMemberLogin.mqrm',
    ENROLMENT_URL = '/OCLoyaltyEnrollment.mqrm',
    GET_TOKEN_URL =  '/getToken.mqrm',
    IMPORT_CONTACT_URL = '/importContact.mqrm',
    UPDATE_CONTACT_URL = '/updateContactsOPT.mqrm',
    LOYALTY_ENQUIRY_URL = '/OCLoyaltyInquiry.mqrm',
    TRANSACTION_HISTORY_URL = '/loyaltyTransactionHistory.mqrm',
    PURCHASE_HISTORY_URL = '/purchaseHistory.mqrm',
    COUPON_HISTORY_URL = '/couponsHistory.mqrm',
    COUPON_CODE_INQUIRY_URL = '/CouponCodeEnquiryRequestOPT.mqrm',
    OFFER_BANNERS_URL = '/api/FCM/GetOfferBanners',
    REGISTER_DEVICE_TOKEN_URL = '/api/FCM/registerInstanceIDToken',
    GET_NOTIFICATIONS_URL = '/api/FCM/getNotifications',
    OCOTPSERVICE_URL = '/OCOTPService.mqrm',
    CMS_COUPON_HISTORY_URL = '/api/FCM/GetOffers',
    GET_COUNTRIES_URL = '/api/FCM/GetCountries',
    GET_CITIES_URL = '/api/FCM/GetCities',
    GET_LOCALITIES_URL = '/api/FCM/GetLocalities',
    GET_BRANDS_URL = '/api/FCM/GetBrands',
    GET_STORE_INFO_URL = '/api/FCM/GetStoreInfo',
    GET_LANG_FEILDS_URL = '/api/FCM/GetLangFields',
    OPT_LANG_URL= '/processLanguage.mqrm',
    GET_STORES_URL = '/api/FCM/GetStores',
    OPT_STORES_INQUIRY_URL = '/storeInquiry.mqrm',
    OPT_NOTIFICATION_URL = '/pushNotification.mqrm ',
    GET_OFFER_DETAIL_URL = '/api/FCM/GetOfferDetails',
    GET_BANNER_URL= '/api/FCM/GetBanners',
    WEB_PORTAL_BRANDING_URL = '/webPortalBranding.mqrm',
    LOG_OUT_URL  = '/loyaltyMemberLogout.mqrm'
}